+++ CHANGELOG.md
@@ -106,7 +106,7 @@ CHANGELOG
   - The following report field names have changed to better represent their values:
       scan_rsync:extra.password renamed to extra.has_password
       scan_elasticsearch:status renamed to http_code
-  - Added `Accessible-HTTP-proxy` (PR#2246 by elsif2).
+  - Added `Accessible-HTTP-proxy` and `Open-HTTP-proxy` (PR#2246 by elsif2).
 
 - `intelmq.bots.parsers.cymru.parser_cap_program`: The parser mapped the hostname into `source.fqdn` which is not allowed by the IntelMQ Data Format. Added a check (PR#2215 by Sebastian Waldbauer, fixes #2169)
 - `intelmq.bots.parsers.generic.parser_csv`: Use RewindableFileHandle to use the original current line for line recovery (PR#2192 by Sebastian Wagner).